move _select_next_nodes to dag_node_operation.py
